Speakerbox/receive/requirements.txt
mplayer.py
paho-mqtt
